Phishing attack and prevention of phishing attacks
 Phishing sites are the biggest threat to users’ 
information since they compromise identity, lead to 
financial loss, and have a negative impact on 
organization’s reputation. This abstract presents an 
innovative protective security model that employs fast 
identification algorithms and behavioral patterns to 
prevent or neutralize new threats. These consist of real 
time website scanning, risk assessment functionality, 
suggestions based on the context currently being used and 
secure modes of encryption. It also has a decentralized 
authentication strategy to minimize data loss and other 
mishaps such as unauthorized access. The performance of 
the solution has been tested and validated to a great extent 
to prove its improvement in the mechanism of detection 
and prevention of phishing threats. These are proactive 
security instruments that are quintessential in the 
preservation of propriety data and users’ confidentiality 
in the electronic setting